:recycle: Rename votingSummary to votingV2
