+++ src/components/votingV2/index.js
