+++ src/components/votingV2/votingSummary.css
