+++ src/components/votingV2/votingSummary.js
